pyproject.toml
[project]
name = "recipe-api"
version = "1.1.0"
description = "Recipe-API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flake8>=7.2",
    "Django==3.2",
    "djangorestframework==3.13",
    "django-filter==23.1",
    "psycopg2-binary==2.9.9",
    "drf-spectacular==0.15",
    "pillow==9.2",
    "uwsgi==2.0.29",
]